utils: move InodeBitmap to dedicate file
Move InodeBitmap to dedicate file, so lib.rs get clear.
